+++ Trove/src/utopia/trove/controller/LocalDatabase.scala
@@ -48,51 +48,59 @@ object LocalDatabase
 	
 	// OTHER	-------------------------------
 	
-	// TODO: Continue
-	/*
-	  * Clears existing data from the database
-	  * @return Success or failure
+	/**
+	  * Drops a versioned database
+	  * @param dbName Name of the managed / targeted database
+	  * @param versionTableName Name of the version table
+	  * @param versionTable A version table access (call by name) (is called only if such a table already exists in db)
+	  * @param exc Implicit execution context
+	  * @param connectionPool Implicit connection pool
+	  * @return Failure if database failed to start or if failed to clear the database
 	  */
-		/*
-	def clear() =  Try {
-		// Sets up the database
-		val configBuilder = DBConfigurationBuilder.newBuilder()
-		configBuilder.setPort(0)
-		configBuilder.setDataDir(("database": Path).absolute.toString)
-		val database = DB.newEmbeddedDB(configBuilder.build()) // May throw
-		
-		// Updates Vault connection settings
-		Connection.modifySettings { _.copy(
-			connectionTarget = configBuilder.getURL(""),
-			defaultDBName = Some("test")) }
-		
-		// Starts the database (may throw)
-		database.start()
-		// Closes the database when program closes
-		CloseHook.registerAction {
-			println("Stopping database")
-			database.stop()
+	def clear(dbName: String, versionTableName: String, versionTable: => Table)
+			 (implicit exc: ExecutionContext, connectionPool: ConnectionPool) =
+	{
+		// Starts the database first if not started already
+		start().flatMap { _ =>
+			connectionPool.tryWith { implicit connection =>
+				// Clears the version table, if available (and if targeting another db)
+				if (connection.existsTable(dbName, versionTableName))
+					DbDatabaseVersions(versionTable).delete()
+				// Drops the target database
+				connection.dropDatabase(dbName)
+			}
 		}
-		
-		// Clears old data
-		connectionPool { _.dropDatabase(Tables.databaseName) }
-	}*/
+	}
 	
 	/**
 	  * Sets up the local database
 	  * @param updateDirectory Directory from which update files are read
-	  * @param versionDbName Name of the database that contains version data
+	  * @param dbName Name of the managed database
 	  * @param versionTableName Name of the table that contains version data
 	  * @param versionTable Version table (call by name)
-	  * @param targetDatabaseName Name of the database targeted by the updates and used by the program.
-	  *                           Empty if not specified (default).
+	  * @param listener Listener to be informed of database setup events
+	  * @param exc Implicit execution context
+	  * @param connectionPool A connection pool for forming database connections
+	  * @return Database setup completion event
+	  */
+	def setupWithListener(updateDirectory: Path, dbName: String, versionTableName: String,
+						  versionTable: => Table)(listener: DatabaseSetupListener)
+						 (implicit exc: ExecutionContext, connectionPool: ConnectionPool) =
+		setup(updateDirectory, dbName, versionTableName, versionTable, Some(listener))
+	
+	/**
+	  * Sets up the local database
+	  * @param updateDirectory Directory from which update files are read
+	  * @param dbName Name of the managed database
+	  * @param versionTableName Name of the table that contains version data
+	  * @param versionTable Version table (call by name)
 	  * @param listener Listener to be informed of database setup events (optional)
 	  * @param exc Implicit execution context
 	  * @param connectionPool A connection pool for forming database connections
 	  * @return Database setup completion event
 	  */
-	def setup(updateDirectory: Path, versionDbName: String, versionTableName: String, versionTable: => Table,
-			  targetDatabaseName: String = "", listener: Option[DatabaseSetupListener] = None)
+	def setup(updateDirectory: Path, dbName: String, versionTableName: String, versionTable: => Table,
+			  listener: Option[DatabaseSetupListener] = None)
 			 (implicit exc: ExecutionContext, connectionPool: ConnectionPool) =
 	{
 		// Table is cached once it has been used once
@@ -100,40 +108,10 @@ object LocalDatabase
 		
 		def fireEvent(event: => DatabaseSetupEvent) = listener.foreach { _.onDatabaseSetupEvent(event) }
 		
-		val result = Try {
-			// If currently processing another request, waits for that to complete first
-			waitUntilNotProcessing()
-			
-			// May skip the database starting process in case there is already a started database active
-			if (_statusPointer.getAndUpdate { s => if (s.notStarted) Starting else Updating }.notStarted)
-			{
-				// Sets up the database
-				val configBuilder = DBConfigurationBuilder.newBuilder()
-				configBuilder.setPort(0)
-				configBuilder.setDataDir(("database": Path).absolute.toString)
-				val database = DB.newEmbeddedDB(configBuilder.build()) // May throw
-				dbPointer.setOne(database)
-				
-				// Updates Vault connection settings
-				Connection.modifySettings { _.copy(
-					connectionTarget = configBuilder.getURL(""),
-					defaultDBName = Some("test")) }
-				
-				fireEvent(DatabaseConfigured)
-				
-				// Starts the database (may throw)
-				database.start()
-				_statusPointer.set(Updating)
-				// Closes the database when program closes
-				CloseHook.registerAction { shutDown() }
-				
-				fireEvent(DatabaseStarted)
-			}
-			
-		}.flatMap { _ =>
+		val result = start(listener).flatMap { _ =>
 			// Checks current database version, and whether database has been configured at all
 			connectionPool.tryWith { implicit connection =>
-				if (connection.existsTable(versionDbName, versionTableName))
+				if (connection.existsTable(dbName, versionTableName))
 					DbDatabaseVersion(table).latest
 				else
 					None
@@ -141,20 +119,45 @@ object LocalDatabase
 				ScanSourceFiles(updateDirectory, currentDbVersion.map { _.number }).flatMap { sources =>
 					// If no update files were available, completes
 					if (sources.isEmpty)
-						Success(SetupSucceeded(currentDbVersion))
+					{
+						// May create the target database first, however
+						Success(
+							if (currentDbVersion.isEmpty)
+								connectionPool.tryWith { _.createDatabase(dbName) } match
+								{
+									case Success(_) => SetupSucceeded(None)
+									case Failure(error) => SetupFailed(error)
+								}
+							else
+								SetupSucceeded(currentDbVersion)
+						)
+					}
 					else
 					{
 						fireEvent(UpdatesFound(sources, currentDbVersion))
 						
 						connectionPool.tryWith { implicit connection =>
-							// Drops or uses the targeted database if necessary
-							if (targetDatabaseName.nonEmpty && currentDbVersion.isDefined)
+							// Creates, recreates and/or uses the targeted database if necessary
+							if (currentDbVersion.isDefined)
 							{
+								// If a full update will be introduced and there already exists a database version,
+								// drops it and recreates it first
 								if (sources.exists { _.fileType == Full })
-									connection.dropDatabase(targetDatabaseName)
+								{
+									// Backs up version data and inserts it to the new database version
+									val versionsAccess = DbDatabaseVersions(table)
+									val versions = versionsAccess.all
+									connection.dropDatabase(dbName, checkIfExists = false)
+									connection.createDatabase(dbName, checkIfExists = false)
+									versionsAccess.insert(versions.map { _.data })
+								}
+								// In case there is an update and db exists already, uses it
 								else
-									connection.dbName = targetDatabaseName
+									connection.dbName = dbName
 							}
+							// If there wasn't any target database before, creates one
+							else
+								connection.createDatabase(dbName)
 							
 							// Imports the source files in order
 							// Keeps track of the version(s) during the loop in order to record the final version
@@ -176,8 +179,7 @@ object LocalDatabase
 							val newVersion = version.map { v => DbDatabaseVersions(table).insert(v) }
 							
 							// Sets up the default database name
-							if (targetDatabaseName.nonEmpty)
-								Connection.modifySettings { _.copy(defaultDBName = Some(targetDatabaseName)) }
+							Connection.modifySettings { _.copy(defaultDBName = Some(dbName)) }
 							
 							// Returns the completion event (without firing it)
 							updateFailure match
@@ -250,6 +252,40 @@ object LocalDatabase
 	{
 		if (isProcessing)
 			_statusPointer.futureWhere { _.isCompleted }.waitFor()
+	}
+	
+	private def start(listener: Option[DatabaseSetupListener] = None)(implicit exc: ExecutionContext) =
+	{
+		Try {
+			// If currently processing another request, waits for that to complete first
+			waitUntilNotProcessing()
+			
+			// May skip the database starting process in case there is already a started database active
+			if (_statusPointer.getAndUpdate { s => if (s.notStarted) Starting else Updating }.notStarted)
+			{
+				// Sets up the database
+				val configBuilder = DBConfigurationBuilder.newBuilder()
+				configBuilder.setPort(0)
+				configBuilder.setDataDir(("database": Path).absolute.toString)
+				val database = DB.newEmbeddedDB(configBuilder.build()) // May throw
+				dbPointer.setOne(database)
+				
+				// Updates Vault connection settings
+				Connection.modifySettings { _.copy(
+					connectionTarget = configBuilder.getURL(""),
+					defaultDBName = Some("test")) }
+				
+				listener.foreach { _.onDatabaseSetupEvent(DatabaseConfigured) }
+				
+				// Starts the database (may throw)
+				database.start()
+				_statusPointer.set(Updating)
+				// Closes the database when program closes
+				CloseHook.registerAction { shutDown() }
+				
+				listener.foreach { _.onDatabaseSetupEvent(DatabaseStarted) }
+			}
+		}
 	}